+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_inputs_outputs.py
@@ -65,6 +65,8 @@ from enum import EnumMeta
 from inspect import Parameter, signature
 from typing import Dict, Iterable, Sequence, Union, overload
 
+from typing_extensions import Literal
+
 from azure.ai.ml._schema.component.input_output import SUPPORTED_PARAM_TYPES
 from azure.ai.ml.constants import AssetTypes
 from azure.ai.ml.constants._component import ComponentParameterTypes, IOConstants
@@ -128,7 +130,7 @@ class Input(_InputOutputBase):  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
-        type: str = "uri_folder",
+        type: Literal["uri_folder"] = "uri_folder",
         path: str = None,
         mode: str = None,
         optional: bool = None,
@@ -160,7 +162,7 @@ class Input(_InputOutputBase):  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
-        type: str = "number",
+        type: Literal["number"] = "number",
         default: float = None,
         min: float = None,
         max: float = None,
@@ -188,7 +190,7 @@ class Input(_InputOutputBase):  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
-        type: str = "integer",
+        type: Literal["integer"] = "integer",
         default: int = None,
         min: int = None,
         max: int = None,
@@ -216,7 +218,7 @@ class Input(_InputOutputBase):  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
-        type: str = "string",
+        type: Literal["string"] = "string",
         default: str = None,
         optional: bool = None,
         description: str = None,
@@ -238,7 +240,7 @@ class Input(_InputOutputBase):  # pylint: disable=too-many-instance-attributes
     def __init__(
         self,
         *,
-        type: str = "boolean",
+        type: Literal["boolean"] = "boolean",
         default: bool = None,
         optional: bool = None,
         description: str = None,
@@ -543,7 +545,14 @@ class Output(_InputOutputBase):
     """
 
     @overload
-    def __init__(self, type="uri_folder", path=None, mode=None, description=None):
+    def __init__(
+        self,
+        *,
+        type: Literal["uri_folder"] = "uri_folder",
+        path=None,
+        mode=None,
+        description=None,
+    ):
         """Define a uri_folder output.
 
         :param type: The type of the data output. Possible values include:
@@ -561,7 +570,7 @@ class Output(_InputOutputBase):
         """
 
     @overload
-    def __init__(self, type="uri_file", path=None, mode=None, description=None):
+    def __init__(self, type: Literal["uri_file"] = "uri_file", path=None, mode=None, description=None):
         """Define a uri_file output.
 
         :param type: The type of the data output. Possible values include: